@@ -2,8 +2,8 @@ name: image_picker_for_web_integration_tests
 publish_to: none
 
 environment:
-  sdk: ^3.1.0
-  flutter: ">=3.13.0"
+  sdk: ^3.3.0
+  flutter: ">=3.19.0"
 
 dependencies:
   flutter:
@@ -11,6 +11,7 @@ dependencies:
   image_picker_for_web:
     path: ../
   image_picker_platform_interface: ^2.8.0
+  web: ^0.5.1
 
 dev_dependencies:
   flutter_test:
