@@ -2,11 +2,11 @@ name: image_picker_for_web
 description: Web platform implementation of image_picker
 repository: https://github.com/flutter/packages/tree/main/packages/image_picker/image_picker_for_web
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+image_picker%22
-version: 3.0.2
+version: 3.0.3
 
 environment:
-  sdk: ^3.1.0
-  flutter: ">=3.13.0"
+  sdk: ^3.3.0
+  flutter: ">=3.19.0"
 
 flutter:
   plugin:
@@ -23,6 +23,7 @@ dependencies:
     sdk: flutter
   image_picker_platform_interface: ^2.9.0
   mime: ^1.0.4
+  web: ^0.5.1
 
 dev_dependencies:
   flutter_test:
