@@ -3,88 +3,107 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:html' as html;
+import 'dart:js_interop';
 import 'dart:math';
 import 'dart:ui';
 
 import 'package:image_picker_platform_interface/image_picker_platform_interface.dart';
+import 'package:web/helpers.dart';
+import 'package:web/web.dart' as web;
 
 import 'image_resizer_utils.dart';
 
 /// Helper class that resizes images.
 class ImageResizer {
   /// Resizes the image if needed.
+  ///
   /// (Does not support gif images)
-  Future<XFile> resizeImageIfNeeded(XFile file, double? maxWidth,
-      double? maxHeight, int? imageQuality) async {
+  Future<XFile> resizeImageIfNeeded(
+    XFile file,
+    double? maxWidth,
+    double? maxHeight,
+    int? imageQuality,
+  ) async {
     if (!imageResizeNeeded(maxWidth, maxHeight, imageQuality) ||
         file.mimeType == 'image/gif') {
       // Implement maxWidth and maxHeight for image/gif
       return file;
     }
     try {
-      final html.ImageElement imageElement = await loadImage(file.path);
-      final html.CanvasElement canvas =
+      final web.HTMLImageElement imageElement = await loadImage(file.path);
+      final web.HTMLCanvasElement canvas =
           resizeImageElement(imageElement, maxWidth, maxHeight);
       final XFile resizedImage =
           await writeCanvasToFile(file, canvas, imageQuality);
-      html.Url.revokeObjectUrl(file.path);
+      web.URL.revokeObjectURL(file.path);
       return resizedImage;
     } catch (e) {
       return file;
     }
   }
 
-  /// function that loads the blobUrl into an imageElement
-  Future<html.ImageElement> loadImage(String blobUrl) {
-    final Completer<html.ImageElement> imageLoadCompleter =
-        Completer<html.ImageElement>();
-    final html.ImageElement imageElement = html.ImageElement();
-    // ignore: unsafe_html
-    imageElement.src = blobUrl;
-
-    imageElement.onLoad.listen((html.Event event) {
-      imageLoadCompleter.complete(imageElement);
-    });
-    imageElement.onError.listen((html.Event event) {
-      const String exception = 'Error while loading image.';
-      imageElement.remove();
-      imageLoadCompleter.completeError(exception);
-    });
+  /// Loads the `blobUrl` into a [web.HTMLImageElement].
+  Future<web.HTMLImageElement> loadImage(String blobUrl) {
+    final Completer<web.HTMLImageElement> imageLoadCompleter =
+        Completer<web.HTMLImageElement>();
+    final web.HTMLImageElement imageElement = web.HTMLImageElement();
+    imageElement
+      // ignore: unsafe_html
+      ..src = blobUrl
+      ..onLoad.listen((web.Event event) {
+        imageLoadCompleter.complete(imageElement);
+      })
+      ..onError.listen((web.Event event) {
+        const String exception = 'Error while loading image.';
+        imageElement.remove();
+        imageLoadCompleter.completeError(exception);
+      });
     return imageLoadCompleter.future;
   }
 
-  /// Draws image to a canvas while resizing the image to fit the [maxWidth],[maxHeight] constraints
-  html.CanvasElement resizeImageElement(
-      html.ImageElement source, double? maxWidth, double? maxHeight) {
+  /// Resizing the image in a canvas to fit the [maxWidth], [maxHeight] constraints.
+  web.HTMLCanvasElement resizeImageElement(
+    web.HTMLImageElement source,
+    double? maxWidth,
+    double? maxHeight,
+  ) {
     final Size newImageSize = calculateSizeOfDownScaledImage(
-        Size(source.width!.toDouble(), source.height!.toDouble()),
+        Size(source.width.toDouble(), source.height.toDouble()),
         maxWidth,
         maxHeight);
-    final html.CanvasElement canvas = html.CanvasElement();
-    canvas.width = newImageSize.width.toInt();
-    canvas.height = newImageSize.height.toInt();
-    final html.CanvasRenderingContext2D context = canvas.context2D;
+    final web.HTMLCanvasElement canvas = web.HTMLCanvasElement()
+      ..width = newImageSize.width.toInt()
+      ..height = newImageSize.height.toInt();
+    final web.CanvasRenderingContext2D context = canvas.context2D;
     if (maxHeight == null && maxWidth == null) {
       context.drawImage(source, 0, 0);
     } else {
-      context.drawImageScaled(source, 0, 0, canvas.width!, canvas.height!);
+      context.drawImageScaled(
+          source, 0, 0, canvas.width.toDouble(), canvas.height.toDouble());
     }
     return canvas;
   }
 
-  /// function that converts a canvas element to Xfile
+  /// Converts a canvas element to [XFile].
+  ///
   /// [imageQuality] is only supported for jpeg and webp images.
   Future<XFile> writeCanvasToFile(
-      XFile originalFile, html.CanvasElement canvas, int? imageQuality) async {
+    XFile originalFile,
+    web.HTMLCanvasElement canvas,
+    int? imageQuality,
+  ) async {
     final double calculatedImageQuality =
         (min(imageQuality ?? 100, 100)) / 100.0;
-    final html.Blob blob =
-        await canvas.toBlob(originalFile.mimeType, calculatedImageQuality);
-    return XFile(html.Url.createObjectUrlFromBlob(blob),
-        mimeType: originalFile.mimeType,
-        name: 'scaled_${originalFile.name}',
-        lastModified: DateTime.now(),
-        length: blob.size);
+    final Completer<XFile> completer = Completer<XFile>();
+    final web.BlobCallback blobCallback = (web.Blob blob) {
+      completer.complete(XFile(web.URL.createObjectURL(blob),
+          mimeType: originalFile.mimeType,
+          name: 'scaled_${originalFile.name}',
+          lastModified: DateTime.now(),
+          length: blob.size));
+    }.toJS;
+    canvas.toBlob(
+        blobCallback, originalFile.mimeType ?? '', calculatedImageQuality.toJS);
+    return completer.future;
   }
 }
